Update
- auto set is_active
- poll password

+++ kod-vote/polls/views.py
@@ -1,21 +1,26 @@
 from django.shortcuts import render, redirect
 from .models import Poll, Poll_Vote, Poll_Choice
 from django.contrib.auth.decorators import login_required, permission_required
-import datetime
 import pygal
-from django.conf import settings
+from django.utils import timezone
+from datetime import datetime
 
 # Create your views here.
 @login_required
 def index_view(request):
-    polls_avaliable = Poll.objects.filter(end_date__gt=datetime.datetime.now()).order_by('-start_date')
-    polls_closed = Poll.objects.filter(end_date__lte=datetime.datetime.now()).order_by('-end_date')
+    polls = Poll.objects.all().order_by('-start_date')
+    available, closed = [], []
+    for poll in polls:
+        if poll.is_available():
+            available.append(poll)
+        else:
+            closed.append(poll)
+
     context = {
-        'polls_avaliable' : polls_avaliable,
-        'polls_closed' : polls_closed
+        'available' : available,
+        'closed' : closed
     }
 
-    context['now'] = datetime.datetime.now()
     return render(request, 'index.html', context)
 
 @login_required
@@ -23,14 +28,14 @@ def create_view(request):
     if request.method == 'POST':
         subject = request.POST.get('subject')
         detail = request.POST.get('detail')
-        
+
         try:
             picture = request.FILES['picture']
         except:
             picture = None
 
-        start_date = datetime.datetime.strptime(request.POST.get('start_date'), '%d/%m/%Y %H:%M')
-        end_date = datetime.datetime.strptime(request.POST.get('end_date'), '%d/%m/%Y %H:%M')
+        start_date = datetime.strptime(request.POST.get('start_date'), '%d/%m/%Y %H:%M')
+        end_date = datetime.strptime(request.POST.get('end_date'), '%d/%m/%Y %H:%M')
         password = request.POST.get('password').strip()
 
         poll = Poll(
@@ -44,32 +49,44 @@ def create_view(request):
 
         if picture != None:
             poll.picture = picture
-        
+
         poll.save()
 
         return redirect('mypolls')
-    time = datetime.datetime.now()
-    return render(request, 'create.html', {'time' : time})
+
+    return render(request, 'create.html')
 
 @login_required
 def detail_view(request, poll_id):
-    context = {}
     poll = Poll.objects.get(id=poll_id)
-    context['poll'] = poll
+    already_vote = False
+    context = {
+        'poll' : poll
+    }
 
     for vote in poll.poll_vote_set.all():
         if vote.vote_by == request.user:
-            context['already_vote'] = True
-            break
+            already_vote = True
+
     choices = poll.poll_choice_set.all()
     context['choices'] = choices
-    context['now'] = datetime.datetime.now()
+
     # pygal
-    chart = pygal.Bar(width=1000, min_scale=1)
-    chart.title = f"ผลลัพธ์ของ : {poll.subject}"
-    for choice in choices:
-       chart.add(choice.subject, choice.poll_vote_set.all().count())
-    context['graph'] = chart.render().decode('utf-8')
+    if not poll.is_available() or already_vote:
+        chart = pygal.Bar(width=1000, min_scale=1)
+        chart.title = f"ผลลัพธ์ของ : {poll.subject}"
+        for choice in choices:
+            chart.add(choice.subject, choice.poll_vote_set.all().count())
+            context['graph'] = chart.render().decode('utf-8')
+
+    context['already_vote'] = already_vote
+
+    passed = False
+
+    if poll.password == request.GET.get('password') or already_vote:
+        passed = True
+
+    context['passed'] = passed
 
     return render(request, 'detail.html', context)
 
@@ -77,6 +94,11 @@ def detail_view(request, poll_id):
 def vote_view(request, choice_id):
     choice = Poll_Choice.objects.get(id=choice_id)
     poll = choice.poll_id
+
+    for vote in request.user.poll_vote_set.all():
+        if vote.poll_id == poll:
+            # already vote
+            return redirect('detail', poll_id=poll.id)
     vote = Poll_Vote.objects.create(
         poll_id=poll,
         choice_id=choice,
@@ -87,6 +109,8 @@ def vote_view(request, choice_id):
 @login_required
 def edit_view(request, poll_id):
     poll = Poll.objects.get(id=poll_id)
+    if request.user != poll.create_by:
+        return redirect('index')
     context = {'poll' : poll}
     if request.method == 'POST':
         subject = request.POST.get('subject')
@@ -100,7 +124,7 @@ def edit_view(request, poll_id):
 
         poll.save()
 
-        context['success'] = 'แก้ไข Poll เรียบร้อยแล้ว'
+        context['success'] = 'แก้ไขกระทู้เรียบร้อยแล้ว'
 
     return render(request, 'edit.html', context)
 
@@ -109,7 +133,8 @@ def close_view(request, poll_id):
     poll = Poll.objects.get(id=poll_id)
 
     if poll.create_by == request.user:
-        poll.end_date = datetime.datetime.now()
+        poll.is_active = False
+        poll.end_date = timezone.now()
         poll.save()
 
     return redirect('mypolls')
@@ -129,31 +154,32 @@ def add_choice_view(request, poll_id):
     context = {
         'poll' : poll
     }
+    if request.user != poll.create_by:
+        return redirect('index')
     if request.method == 'POST':
         try:
             image = request.FILES['picture']
         except:
             image = None
-
         choice = Poll_Choice(
             subject = request.POST.get('subject'),
             poll_id=poll
         )
-
         if image != None:
             choice.image = image
-        
         choice.save()
-
         return redirect('edit', poll_id=poll.id)
+
     return render(request, 'add_choice.html', context)
 
 @login_required
 def choice_delete_view(request, choice_id):
     choice = Poll_Choice.objects.get(id=choice_id)
+    poll = choice.poll_id
 
-    poll_id = choice.poll_id.id
+    if request.user != poll.create_by:
+        return redirect('index')
 
     choice.delete()
 
-    return redirect('edit', poll_id=poll_id)
+    return redirect('edit', poll_id=poll.id)
